requirements.txt
Pillow
pypdf
requests
